+++ pyo3-derive-backend/src/pymethod.rs
@@ -445,7 +445,6 @@ fn impl_arg_params_(spec: &FnSpec<'_>, body: TokenStream, into_result: TokenStre
     let num_normal_params = params.len();
     // create array of arguments, and then parse
     quote! {
-        use pyo3::ObjectProtocol;
         const PARAMS: &'static [pyo3::derive_utils::ParamDescription] = &[
             #(#params),*
         ];