+++ src/types/list.rs
@@ -222,7 +222,6 @@ where
 #[cfg(test)]
 mod test {
     use crate::instance::AsPyRef;
-    use crate::objectprotocol::ObjectProtocol;
     use crate::types::PyList;
     use crate::Python;
     use crate::{IntoPy, PyObject, PyTryFrom, ToPyObject};